1.Create module for Product Library v2: 'snap-product-library-v2' in snap engine and 'snap-product-library-ui-v2' in snap desktop. 2. Move classes of the Product Library v2 in the new modules.
