+++ pom.xml
@@ -42,6 +42,7 @@
         <module>snap-land-cover-ui</module>
         <module>snap-pixel-extraction-ui</module>
         <module>snap-product-library-ui</module>
+        <module>snap-product-library-ui-v2</module>
         <module>snap-raster-ui</module>
         <module>snap-ui</module>
         <module>snap-unmix-ui</module>
