+++ snap-application/pom.xml
@@ -113,6 +113,11 @@
             <version>${project.version}</version>
         </dependency>
         <dependency>
+            <groupId>org.esa.snap</groupId>
+            <artifactId>snap-product-library-ui-v2</artifactId>
+            <version>${project.version}</version>
+        </dependency>
+        <dependency>
             <groupId>${project.groupId}</groupId>
             <artifactId>snap-raster-ui</artifactId>
             <version>${project.version}</version>
