+++ snap-product-library-ui-v2/pom.xml
@@ -1,23 +1,9 @@
 <?xml version="1.0"?>
-<!--
- * Copyright (C) 2015 by Array Systems Computing Inc. http://www.array.ca
- *
- * This program is free software; you can redistribute it and/or modify it
- * under the terms of the GNU General Public License as published by the Free
- * Software Foundation; either version 3 of the License, or (at your option)
- * any later version.
- * This program is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or
- * FITNESS FOR A PARTICULAR PURPOSE. See the GNU General Public License for
- * more details.
- *
- * You should have received a copy of the GNU General Public License along
- * with this program; if not, see http://www.gnu.org/licenses/
-  -->
-<project
-        xmlns="http://maven.apache.org/POM/4.0.0"
-        xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-        xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
+
     <modelVersion>4.0.0</modelVersion>
 
     <parent>
@@ -26,8 +12,9 @@
         <version>8.0.0-SNAPSHOT</version>
     </parent>
 
-    <name>SNAP Product Library UI</name>
-    <artifactId>snap-product-library-ui</artifactId>
+    <artifactId>snap-product-library-ui-v2</artifactId>
+
+    <name>SNAP Product Library UI v2</name>
 
     <packaging>nbm</packaging>
 
@@ -79,11 +66,6 @@
         <dependency>
             <groupId>org.esa.snap</groupId>
             <artifactId>snap-product-library-v2</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.esa.snap</groupId>
-            <artifactId>snap-product-library-ui-v2</artifactId>
             <version>${project.version}</version>
         </dependency>
         <dependency>
