+++ snap-product-library-ui-v2/src/main/java/org/esa/snap/product/library/ui/v2/repository/local/LocalProgressStatus.java
