+++ snap-product-library-ui-v2/src/main/java/org/esa/snap/product/library/ui/v2/worldwind/Rectangle3DSelection.java
@@ -5,7 +5,6 @@ import gov.nasa.worldwind.geom.Sector;
 import gov.nasa.worldwind.layers.LayerList;
 import gov.nasa.worldwindx.examples.util.SectorSelector;
 
-import java.awt.Cursor;
 import java.awt.geom.Rectangle2D;
 
 /**
