+++ snap-product-library-ui/src/main/java/org/esa/snap/product/library/ui/v2/repository/AbstractRepositoryProductPanel.java
@@ -3,7 +3,6 @@ package org.esa.snap.product.library.ui.v2.repository;
 import org.apache.commons.lang.StringUtils;
 import org.esa.snap.product.library.ui.v2.ComponentDimension;
 import org.esa.snap.product.library.ui.v2.RepositoryProductPanelBackground;
-import org.esa.snap.product.library.ui.v2.repository.output.OutputProductListModel;
 import org.esa.snap.product.library.ui.v2.repository.output.OutputProductResults;
 import org.esa.snap.remote.products.repository.Attribute;
 import org.esa.snap.remote.products.repository.RepositoryProduct;
